+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Breaking Changes:
 
 - Use datetime2 for MSSQL datetime + timestamp types. This change is incompatible with MSSQL older than 2008 #2757
+- Introduced support for specifying multiple directories for the Migrator #2735
 - Knex.VERSION() method was removed, run "require('knex/package').version" instead #2776
 
 # 0.15.2 - 19 Jul, 2018